Fix EZP-24157: eZFlow Block names swapped when moving blocks in layout
Change the behaviour of the up and down buttons to be same as when the user
drags and drops a block in the list. In addition, this patch adds the detection
of the ezautosave to use it to store the state of the blocks.

(cherry picked from commit cb7689f7ff47b3319cf9534852df7a92a72c6405)